requirements-2.txt
flask>=3.0.0
gunicorn>=21.2.0
pytesseract>=0.3.13
opencv-python-headless>=4.9.0.80
sympy>=1.12
Pillow>=10.2.0
numpy>=1.26.0
werkzeug>=3.0.0
